requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cffi==2.0.0
click==8.1.8
colorama==0.4.6 ; sys_platform == "win32" or platform_system == "Windows"
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1 ; python_version < "3.11"
fastapi==0.122.1
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pwdlib==0.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyjwt==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pyyaml==6.0.3
rsa==4.2
six==1.17.0
starlette==0.49.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
uvloop==0.22.1 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.1
websockets==15.0.1
sqlalchemy
aiomysql
pymysql
torch
transformers
numpy
qdrant-client